pyproject.toml
[tool.poetry]
name = "practice"
version = "0.1.0"
description = ""
authors = ["강석영 <ksge1124@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.1.2"
psycopg2-binary = "^2.9.5"
djangorestframework = "^3.14.0"
django-rest-framework = "^0.1.0"


[tool.poetry.group.dev.dependencies]
flake8 = "^5.0.4"
mypy = "^0.982"
black = "^22.10.0"
django-stubs = "^1.12.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
